restructured, please repair linkage error from here
