--- src/debug_util.h
@@ -1,11 +1,0 @@
-#pragma once
-
-#include <iostream>
-
-namespace debug {
-	[[deprecated]] void read() {
-		char c;
-		std::cin >> c;
-	}
-
-}
