+++ src/logger.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "spdlog/logger.h"
+
+spdlog::logger& standard_logger();
+
+void init_logger();
