+++ src/otp.cpp
@@ -2,12 +2,11 @@
 
 
 #include "internal_error.h"
-#include "debug_util.h"
+#include "global_data.h"
+#include "logger.h"
 
 #include "nlohmann/json.hpp"
 
-#include "spdlog/spdlog.h"
-#include "spdlog/sinks/ostream_sink.h"
 
 #include <filesystem>
 #include <iostream>
@@ -77,28 +76,7 @@ otp [-D | -F | -DF ] [action_name] [action params]
 
 */
 
-static const std::string STANDARD_LOGGER_NAME{ "otplog" };
 
-spdlog::logger& standard_logger() {
-	auto raw_ptr = spdlog::get(STANDARD_LOGGER_NAME).get();
-	internal_error::assert_true(raw_ptr != nullptr, "The standard logger has not been registered.");
-	return *raw_ptr;
-}
-
-struct const_strings {
-
-	static constexpr std::string_view CONFIG_FILE_STANDARD_NAME{ "config.json" };
-
-	static constexpr std::string_view EMPTY_CONFIG{ R"xxx({
-   pads: []
-   config: {
-      auto_delete_used_keyfiles: false
-   }
-}
-)xxx" };
-
-
-};
 
 class action_error : public std::exception {
 	std::string message;
@@ -119,106 +97,6 @@ public:
 };
 
 
-class bad_argument : public std::exception {
-	std::string message;
-public:
-	static void assert_true(bool condition, const action_error& exception) {
-		if (!condition) throw exception;
-	}
-
-	bad_argument(const std::string& message) : message(message) {}
-	bad_argument(const char* message) : bad_argument(std::string(message)) {}
-	bad_argument(const bad_argument&) = default;
-
-
-	const char* what() const noexcept override {
-		return message.c_str();
-	}
-};
-
-/**
-* @brief Contains all data and structure instances appearing global for OTP
-*/
-class global_data {
-
-	std::filesystem::path _config_file_path; // absolute path to config file, weakly_cannonical, i.e. containing no .. or .
-
-	std::vector<std::string> _action_command_sequence; // commands regarding the action to be executed (no executable path, no -C option anymore)
-
-	std::filesystem::path executable_path; // path to the executable which is running
-
-	std::filesystem::path working_directory; // current working directory
-
-	inline std::filesystem::path pad_root_directory() { return  _config_file_path.parent_path(); } // absolute path to the root directory of the pad, which is the directory where the config.json is located
-
-public:
-
-	const std::vector<std::string> application_start_arguments; // all arguments as passed via the commandline
-
-	inline const std::filesystem::path& config_file_path() const { return _config_file_path; }
-
-	inline bool exists_config_file() const { return std::filesystem::exists(config_file_path()); }
-
-private:
-
-	std::vector<std::string> decode_command_line_input(int argc, char** argv) {
-		auto result{ std::vector<std::string>(argc) };
-		for (decltype(argc) i = 0; i < argc; ++i)
-			result[i] = argv[i];
-		return result;
-	}
-
-public:
-
-	global_data(int argc, char** argv) : application_start_arguments(decode_command_line_input(argc, argv)) {
-		standard_logger().debug("Initializing global data...");
-		try {
-			executable_path = application_start_arguments[0];
-			working_directory = std::filesystem::current_path();
-		}
-		catch (...) {
-			throw internal_error("Error when initializing working directories.");
-		}
-		standard_logger().debug(std::string("path to executable: ") + executable_path.generic_string());
-		standard_logger().debug(std::string("working directory: ") + working_directory.generic_string());
-
-		unsigned int skip_arg_counter{ 1 };
-
-		if ((application_start_arguments.size()>1) && (application_start_arguments[1] == "-C")) {
-			// use a config file located at a custom directory with a custom name.
-			std::filesystem::path given_path = application_start_arguments[2];
-			_config_file_path = given_path.has_root_path() ? 
-				given_path : 
-				working_directory / given_path ;
-			try {
-				_config_file_path = std::filesystem::weakly_canonical(_config_file_path);
-			}
-			catch (const std::filesystem::filesystem_error& e) {
-				standard_logger().error(std::string("Error on processing config file path: ") + e.what());
-				throw bad_argument("Given config file path might be ill-formed.");
-			}
-			skip_arg_counter += 2;
-			standard_logger().info(std::string("Set a custom config file path: ") + _config_file_path.generic_string());
-		}
-		else {
-			_config_file_path = working_directory / const_strings::CONFIG_FILE_STANDARD_NAME;
-			standard_logger().debug(std::string("Implicitly use standard config file path: ") + _config_file_path.generic_string());
-		}
-
-		throw "move on here";
-
-		internal_error::assert_true(application_start_arguments.cend() - application_start_arguments.cbegin() <= skip_arg_counter,
-			"Skipping arguments on global data initialisation failed.");
-		for (auto it = application_start_arguments.cbegin() + skip_arg_counter; it != application_start_arguments.cend(); ++it) {
-			_action_command_sequence.push_back(*it);
-		}
-	}
-
-	std::filesystem::path custom_working_directory() const {
-		return std::filesystem::path();
-	}
-
-};
 
 
 
@@ -267,7 +145,7 @@ protected:
 		action_error::assert_true(config_json.good(), "Could not create config.json.");
 		config_json << const_strings::EMPTY_CONFIG;
 		config_json.close();
-		std::cout << "Successfully created a fresh config.json:\n\t" << data->custom_working_directory() << '\n';
+		std::cout << "Successfully created a fresh config.json:\n\t" << data->config_file_path() << '\n';
 	}
 
 };
@@ -304,12 +182,12 @@ int cli(int argc, char** argv) {
 		all_actions.emplace_front(new otp_init());
 
 		// check for actions matching the input:
-		std::list<action*> matching_actions;
-		std::for_each(all_actions.cbegin(), all_actions.cend(), [&](const auto& p_action) { if (p_action->match()) matching_actions.emplace_back(p_action.get()); });
+		std::forward_list<action*> matching_actions;
+		std::for_each(all_actions.cbegin(), all_actions.cend(), [&](const auto& p_action) { if (p_action->match()) matching_actions.push_front(p_action.get()); });
 
 		// Check for missing or ambiguous actions:
-		action_error::assert_true(matching_actions.size() > 0, "Invalid input. There is no otp action matching your input.");
-		internal_error::assert_true(matching_actions.size() < 2,
+		action_error::assert_true(!matching_actions.empty(), "Invalid input. There is no otp action matching your input.");
+		internal_error::assert_true(std::next(matching_actions.cbegin()) == matching_actions.cend(),
 			std::string("More than one action matched for input sequence which is:\n  ")
 			+ std::accumulate(data.application_start_arguments.cbegin(), data.application_start_arguments.cend(), std::string(), [](const auto& l, const auto& r) { return l + " " + r; })
 			+ "\nThe matching actions are {"
@@ -350,11 +228,7 @@ int cli(int argc, char** argv) {
 
 int main(int argc, char** argv)
 {
-	// set up logger
-	auto sink_std_cout = std::make_shared<spdlog::sinks::ostream_sink_mt>(std::cout);
-	auto standard_logger = std::make_shared<spdlog::logger>(STANDARD_LOGGER_NAME, sink_std_cout);
-	standard_logger->set_level(spdlog::level::debug);
-	spdlog::register_logger(standard_logger);
+	init_logger();
 
 	return cli(argc, argv);
